backend/requirements.txt
Django>=4.1.4,<4.2
djangorestframework>=3.14.0,<3.15
channels>=4.0.0,<4.1
Pillow>=9.4.0,<9.5
daphne>=4.0.0,<4.1